requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==23.1.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
numpy==1.24.3
outcome==1.2.0
pandas==2.0.1
psutil==5.9.5
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.14
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
wsproto==1.2.0
